% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP reduction of an NMF model}
\usage{
plot_umap(nmf_model, ...)
}
\arguments{
\item{nmf_model}{object of class \code{\link{RcppML::nmf}}.}

\item{...}{arguments to \code{\link{uwot::umap}}}
}
\value{
ggplot2 object of UMAP coordinates with basic styling, you can use other ggplot2 parameters to style the plot further
}
\description{
Plots the sample embeddings matrix of an NMF model, grouping based on the colnames.
}
